Add support for PSSG image format
PSSG is actually a 3d model format but so far we can only extract the
images from it. It has been tested with the PS3 version of Hyperdimension
Neptunia. Probably over time some more pixel format variants will crop
up that need to be handled, but this is a first step.